requirements.txt
# OpenAI API客户端
openai>=1.0.0

# 深度学习框架
torch>=1.12.0
torch-geometric>=2.0.0

# 自然语言处理
transformers>=4.20.0
sentence-transformers>=2.2.0

# 图处理
networkx>=2.8.0

# 数据处理
numpy>=1.21.0
pandas>=1.4.0

# 环境变量管理
python-dotenv>=0.19.0

# JSONL文件处理
jsonlines>=3.1.0

# 异步支持
asyncio

# 正则表达式
re

# 系统操作
os
json
time